client/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
bidict==0.23.1
bitarray==3.4.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
ConfigArgParse==1.7.1
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
frozenlist==1.7.0
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
h11==0.16.0
hexbytes==1.3.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.37.11
locust-cloud==1.24.2
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.6.3
numpy==2.3.1
parsimonious==0.10.0
platformdirs==4.3.8
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyNaCl==1.5.0
python-engineio==4.12.2
python-socketio==5.13.0
pyunormalize==16.0.0
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
rlp==4.1.0
setuptools==80.9.0
simple-websocket==1.1.0
toolz==1.0.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
web3==7.12.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
zope.event==5.1
zope.interface==7.2
